moved exafs energy range builder inside exafs.py
